+++ ai-logic/firebase-ai/src/main/kotlin/com/google/firebase/ai/type/Exceptions.kt
@@ -187,7 +187,7 @@ internal constructor(location: String, cause: Throwable? = null) :
 /**
  * The service is not enabled for this Firebase project. Learn how to enable the required services
  * in the
- * [Firebase documentation.](https://firebase.google.com/docs/vertex-ai/faq-and-troubleshooting#required-apis)
+ * [Firebase documentation.](https://firebase.google.com/docs/ai-logic/faq-and-troubleshooting#required-apis)
  */
 public class ServiceDisabledException
 internal constructor(message: String, cause: Throwable? = null) :
@@ -195,7 +195,7 @@ internal constructor(message: String, cause: Throwable? = null) :
 
 /**
  * The request has hit a quota limit. Learn more about quotas in the
- * [Firebase documentation.](https://firebase.google.com/docs/vertex-ai/quotas)
+ * [Firebase documentation.](https://firebase.google.com/docs/ai-logic/quotas)
  */
 public class QuotaExceededException
 internal constructor(message: String, cause: Throwable? = null) :