+++ ai-logic/firebase-ai-ondevice-interop/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [feature] Added support for on-device structured output generation using `generateObject` (#8395)
+
 # 16.0.0-beta03
 
 - [feature] Added `modelVersion` property to `GenerateContentResponse` (#8227)