+++ ai-logic/firebase-ai-ondevice/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- [feature] Added support for on-device structured output generation using `generateObject` (#8395)
+
 # 16.0.0-beta04
 
 - [feature] Added `getOnDeviceModelName` to `GenerativeModel` (#8247)